Add Connectivity.frontWalk
The Connectivity class now has an frontFactory and FrontWalk, much
like the Adjacency class. Thus you can now do a direct walk on the
Connectivity, without needing to create the Adjacency table. This may
be interesting for large models, where the Adjacency table spends a lot
of memory.

For the practical use, the Mesh.frontWalk method now has a parameter
optim_mem = False/True. The default False will walk using Adjacency
as before. With optim=True, the walk will be done using just the
Connectivity table (a newly created table if level > 0, but might still
use less memory.)

Some tests will be done later to evaluate memory usage and time,
to compare efficiency of both methods.

+++ pyformex/adjacency.py
@@ -373,7 +373,7 @@ class Adjacency(ndarray):
         [ 0  1  1  2 -1]
         [0 1 1 2 4]
         """
-        p = -ones((self.nelems()),dtype=int)
+        p = -ones((self.nelems()),dtype=Int)
         if self.nelems() <= 0:
             return
